Rename alignedmapping→aligned_mapping
