+++ anndata/_core/anndata.py
@@ -26,7 +26,7 @@ from scipy.sparse import issparse
 from .raw import Raw
 from .index import _normalize_indices, _subset, Index, Index1D
 from .file_backing import AnnDataFileManager
-from .alignedmapping import (
+from .aligned_mapping import (
     AxisArrays,
     AxisArraysView,
     PairwiseArrays,
